square payments

+++ app/Http/Controllers/PaymentController.php
@@ -908,7 +908,7 @@ class PaymentController extends Controller
 
         $payment_body = [
             'amount_money' => [
-                'amount' => (int) $request->amount * 100,
+                'amount' => (float) $request->amount * 100,
                 'currency' => 'USD',
             ],
             'idempotency_key' => (string) Str::uuid(),
